fixes
* fixed a crash with fire overlay
* fixed textures pixel accuracy
* fixed fire overlay rendering on the gui screen

+++ src/main/java/net/blancworks/figura/FiguraMod.java
@@ -67,7 +67,7 @@ public class FiguraMod implements ClientModInitializer {
     public static KeyBinding actionWheel;
     public static KeyBinding playerPopup;
 
-    public static long ticksElapsed;
+    public static int ticksElapsed;
 
     //Loading
